backend/pyproject.toml
[tool.poetry]
name = "url_shortener"
version = "0.1.0"
description = ""
authors = ["Edward Fu <edwardfu850808@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
graphene-django = "^2.15.0"
django-graphql-playground = "^1.0"
graphene-federation = "^0.2.0"
Django = "^3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
